4205 - Move variablesToMeasures
