+++ src/meta/measurement/measures/index.ts
@@ -1,7 +1,8 @@
 import { TableName } from 'meta/assessment/table'
 import { VariableName } from 'meta/assessment/variable'
 import { MeasureName } from 'meta/measurement/measure/measure'
-import { measureToVariables, variableToMeasures } from 'meta/measurement/measure/variablesToMeasures'
+
+import { measureToVariables, variableToMeasures } from './variablesToMeasures'
 
 const measureNameToVariableName = (measureName: MeasureName): VariableName => {
   return measureToVariables[measureName] ?? measureName
@@ -10,6 +11,7 @@ const measureNameToVariableName = (measureName: MeasureName): VariableName => {
 const variableNameToMeasureName = (tableName: TableName, variableName: VariableName): MeasureName | VariableName => {
   return variableToMeasures[tableName]?.[variableName] ?? variableName
 }
+
 export const Measures = {
   measureNameToVariableName,
   variableNameToMeasureName,
